test: initial tracing-cov-mark crate implementation
