+++ crates/network-scanner-net/Cargo.toml
@@ -17,6 +17,7 @@ tokio-stream = "0.1.14"
 tracing = "0.1.40"
 
 [dev-dependencies]
+tracing-cov-mark = { path = "../tracing-cov-mark" }
 tracing-subscriber = "0.3.18"
 tokio = { version = "1.36.0", features = [
     "rt",
