+++ crates/network-scanner-net/src/tests.rs
@@ -179,17 +179,21 @@ async fn work_with_tokio_tcp() -> anyhow::Result<()> {
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
 pub async fn drop_runtime() -> anyhow::Result<()> {
+    let (cov, _guard) = tracing_cov_mark::init_cov_mark();
     let kill_server = Arc::new(AtomicBool::new(false));
     let (addr, handle) = local_tcp_server(kill_server.clone()).await?;
+
     tracing_subscriber::fmt::SubscriberBuilder::default()
         .with_max_level(tracing::Level::TRACE)
         .with_thread_names(true)
         .init();
+
     {
-        let async_runtime = crate::runtime::Socket2Runtime::new(None)?;
+        let runtime = crate::runtime::Socket2Runtime::new(None)?;
+
         {
-            let good_socket = async_runtime.new_socket(socket2::Domain::IPV4, socket2::Type::STREAM, None)?;
-            let bad_socket = async_runtime.new_socket(socket2::Domain::IPV4, socket2::Type::STREAM, None)?;
+            let good_socket = runtime.new_socket(socket2::Domain::IPV4, socket2::Type::STREAM, None)?;
+            let bad_socket = runtime.new_socket(socket2::Domain::IPV4, socket2::Type::STREAM, None)?;
 
             tracing::info!("good_socket: {:?}", good_socket);
             tracing::info!("bad_socket: {:?}", bad_socket);
@@ -208,7 +212,7 @@ pub async fn drop_runtime() -> anyhow::Result<()> {
                 );
 
             let (a, b) = tokio::join!(handle, handle2);
-            // remove the outer error from tokio task
+            // Remove the outer error from tokio task.
             let a = a?;
             let b = b?;
             tracing::info!("should connect: {:?}", &a);
@@ -216,13 +220,17 @@ pub async fn drop_runtime() -> anyhow::Result<()> {
             assert!(a.is_ok());
             assert!(b.is_err());
         }
-        tracing::info!("runtime arc count: {}", Arc::strong_count(&async_runtime));
 
-        assert!(Arc::strong_count(&async_runtime) == 1);
+        tracing::info!("runtime arc count: {}", Arc::strong_count(&runtime));
+        assert!(Arc::strong_count(&runtime) == 1);
     }
+
     tracing::info!("runtime should be dropped here");
+    cov.assert_mark("socket2_runtime_drop");
+
     kill_server.store(true, std::sync::atomic::Ordering::Relaxed);
     handle.abort();
+
     Ok(())
 }
