+++ devolutions-gateway/Cargo.toml
@@ -107,3 +107,4 @@ proptest = "1.3"
 rstest = "0.18"
 devolutions-gateway-generators = { path = "../crates/devolutions-gateway-generators" }
 http-body-util = "0.1"
+tracing-cov-mark = { path = "../crates/tracing-cov-mark" }
