+++ devolutions-gateway/tests/webapp.rs
@@ -1,4 +1,4 @@
-use std::{net::SocketAddr, sync::Arc};
+use std::net::SocketAddr;
 
 use anyhow::Context as _;
 use axum::{
@@ -8,12 +8,11 @@ use axum::{
 };
 use axum_extra::headers::{self, HeaderMapExt as _};
 use http_body_util::BodyExt as _;
-use parking_lot::Mutex;
 use serde_json::json;
 use tap::prelude::*;
 use tower::Service as _;
 use tower::ServiceExt as _;
-use tracing::span;
+use tracing_cov_mark::init_cov_mark;
 
 const CONFIG: &str = r#"{
     "ProvisionerPublicKeyData": {
@@ -51,91 +50,6 @@ fn initialize_conf() {
         std::fs::write(users_txt_file, CONTENTS.as_bytes()).unwrap();
         std::env::set_var("DGATEWAY_CONFIG_PATH", std::env!("CARGO_TARGET_TMPDIR"));
     });
-}
-
-// TODO(@CBenoit): move that to another crate for usage in other tests?
-#[derive(Clone, Debug)]
-struct CovMarkSubscriber {
-    records: Arc<Mutex<Vec<String>>>,
-}
-
-#[derive(Clone, Debug)]
-struct CovMarkHandle {
-    records: Arc<Mutex<Vec<String>>>,
-}
-
-#[derive(Clone, Debug)]
-struct CovMarkVisitor {
-    records: Arc<Mutex<Vec<String>>>,
-}
-
-fn init_cov_mark() -> (CovMarkHandle, tracing::subscriber::DefaultGuard) {
-    let subscriber = CovMarkSubscriber::new();
-    let cov_handle = subscriber.handle();
-    let default_guard = tracing::subscriber::set_default(subscriber);
-    (cov_handle, default_guard)
-}
-
-impl CovMarkSubscriber {
-    pub fn new() -> Self {
-        Self {
-            records: Arc::new(Mutex::new(Vec::new())),
-        }
-    }
-
-    pub fn handle(&self) -> CovMarkHandle {
-        CovMarkHandle {
-            records: self.records.clone(),
-        }
-    }
-}
-
-impl tracing::Subscriber for CovMarkSubscriber {
-    fn enabled(&self, _metadata: &tracing::Metadata<'_>) -> bool {
-        true
-    }
-
-    fn new_span(&self, _span: &span::Attributes<'_>) -> span::Id {
-        span::Id::from_u64(1)
-    }
-
-    fn record(&self, _span: &span::Id, _values: &span::Record<'_>) {}
-
-    fn record_follows_from(&self, _span: &span::Id, _follows: &span::Id) {}
-
-    fn event(&self, event: &tracing::Event<'_>) {
-        let mut visitor = CovMarkVisitor {
-            records: self.records.clone(),
-        };
-        event.record(&mut visitor);
-    }
-
-    fn enter(&self, _span: &span::Id) {}
-
-    fn exit(&self, _span: &span::Id) {}
-}
-
-impl CovMarkHandle {
-    #[track_caller]
-    fn assert_mark(&self, covmark: &str) {
-        let mut guard = self.records.lock();
-        let idx = guard
-            .iter()
-            .enumerate()
-            .find_map(|(idx, mark)| (mark.as_str() == covmark).then_some(idx))
-            .expect("coverage marker not emitted");
-        guard.remove(idx);
-    }
-}
-
-impl tracing::field::Visit for CovMarkVisitor {
-    fn record_debug(&mut self, _field: &tracing::field::Field, _value: &dyn std::fmt::Debug) {}
-
-    fn record_str(&mut self, field: &tracing::field::Field, value: &str) {
-        if field.name() == "covmark" {
-            self.records.lock().push(value.to_owned());
-        }
-    }
 }
 
 #[tokio::test]
